cmd/gazelle: split fix/update functionality into new file
* Moved everything related to fix and update commands into fix-update.go
* Renamed main.go to gazelle.go
* Adjusted flag parsing.
