+++ cmd/gazelle/fix_test.go
@@ -23,7 +23,6 @@ import (
 	"testing"
 
 	bf "github.com/bazelbuild/buildtools/build"
-	"github.com/bazelbuild/bazel-gazelle/config"
 )
 
 func TestMain(m *testing.M) {
@@ -32,16 +31,12 @@ func TestMain(m *testing.M) {
 	os.Exit(m.Run())
 }
 
-func defaultConfig(dir string) *config.Config {
-	c := &config.Config{
-		Dirs:                []string{dir},
-		RepoRoot:            dir,
-		GoPrefix:            "example.com/repo",
-		GenericTags:         config.BuildTags{},
-		ValidBuildFileNames: config.DefaultValidBuildFileNames,
+func defaultArgs(dir string) []string {
+	return []string{
+		"-repo_root", dir,
+		"-go_prefix", "example.com/repo",
+		dir,
 	}
-	c.PreprocessTags()
-	return c
 }
 
 func TestFixFile(t *testing.T) {
@@ -68,7 +63,10 @@ func TestFixFile(t *testing.T) {
 		},
 	}
 
-	c := defaultConfig(dir)
+	c, _, err := newFixUpdateConfiguration(defaultArgs(dir))
+	if err != nil {
+		t.Fatalf("newFixUpdateConfiguration faied with %v; want success", err)
+	}
 	if err := fixFile(c, stubFile); err != nil {
 		t.Errorf("fixFile(%#v) failed with %v; want success", stubFile, err)
 		return
@@ -99,8 +97,7 @@ func TestCreateFile(t *testing.T) {
 	}
 
 	// Check that Gazelle creates a new file named "BUILD.bazel".
-	c := defaultConfig(dir)
-	run(c, updateCmd, fixFile)
+	run(defaultArgs(dir))
 
 	buildFile := filepath.Join(dir, "BUILD.bazel")
 	if _, err = os.Stat(buildFile); err != nil {
@@ -128,8 +125,7 @@ func TestUpdateFile(t *testing.T) {
 	}
 
 	// Check that Gazelle updates the BUILD file in place.
-	c := defaultConfig(dir)
-	run(c, updateCmd, fixFile)
+	run(defaultArgs(dir))
 	if st, err := os.Stat(buildFile); err != nil {
 		t.Errorf("could not stat BUILD: %v", err)
 	} else if st.Size() == 0 {
